chore(Table): combine two implementation into single file
