+++ playground/Table.html
@@ -36,63 +36,90 @@
 
   <body>
     <div class="container">
-      <sgds-table-2>
-        <sgds-table-head>
-          <sgds-table-row>
-            <sgds-table-cell>#</sgds-table-cell>
-            <sgds-table-cell>First name</sgds-table-cell>
-            <sgds-table-cell>Last name</sgds-table-cell>
-            <sgds-table-cell>Username</sgds-table-cell>
-            <sgds-table-cell>Action</sgds-table-cell>
-          </sgds-table-row>
-        </sgds-table-head>
+      <sgds-table>
+        <sgds-table-row>
+          <sgds-table-head>#</sgds-table-head>
+          <sgds-table-head>First name</sgds-table-head>
+          <sgds-table-head>Last name</sgds-table-head>
+          <sgds-table-head>Username</sgds-table-head>
+          <sgds-table-head>Action</sgds-table-head>
+        </sgds-table-row>
 
-        <sgds-table-body>
-          <sgds-table-row>
-            <sgds-table-cell>1</sgds-table-cell>
-            <sgds-table-cell>John</sgds-table-cell>
-            <sgds-table-cell>Doe</sgds-table-cell>
-            <sgds-table-cell>
-              <sgds-link>
-                <a href="#">@johndoe</a>
-              </sgds-link>
-            </sgds-table-cell>
-            <sgds-table-cell>
-              <sgds-icon-button name="three-dots-vertical"></sgds-icon-button>
-            </sgds-table-cell>
-          </sgds-table-row>
-          <sgds-table-row>
-            <sgds-table-cell>2</sgds-table-cell>
-            <sgds-table-cell>Jane</sgds-table-cell>
-            <sgds-table-cell>Doe</sgds-table-cell>
-            <sgds-table-cell>
-              <sgds-link>
-                <a href="#">@janedoe</a>
-              </sgds-link>
-            </sgds-table-cell>
-            <sgds-table-cell>-</sgds-table-cell>
-          </sgds-table-row>
-
-          <sgds-table-row>
-            <sgds-table-cell>3</sgds-table-cell>
-            <sgds-table-cell>Bob</sgds-table-cell>
-            <sgds-table-cell>Smith</sgds-table-cell>
-            <sgds-table-cell>
-              <sgds-link>
-                <a href="#">@bobsmith</a>
-              </sgds-link>
-            </sgds-table-cell>
-            <sgds-table-cell>
-              <sgds-badge outlined> active </sgds-badge>
-            </sgds-table-cell>
-          </sgds-table-row>
-        </sgds-table-body>
-      </sgds-table-2>
+        <sgds-table-row>
+          <sgds-table-cell>1</sgds-table-cell>
+          <sgds-table-cell>John</sgds-table-cell>
+          <sgds-table-cell>Doe</sgds-table-cell>
+          <sgds-table-cell>
+            <sgds-link>
+              <a href="#">@johndoe</a>
+            </sgds-link>
+          </sgds-table-cell>
+          <sgds-table-cell>
+            <sgds-icon-button name="three-dots-vertical"></sgds-icon-button>
+          </sgds-table-cell>
+        </sgds-table-row>
+        <sgds-table-row>
+          <sgds-table-cell>2</sgds-table-cell>
+          <sgds-table-cell>Jane</sgds-table-cell>
+          <sgds-table-cell>Doe</sgds-table-cell>
+          <sgds-table-cell>
+            <sgds-link>
+              <a href="#">@janedoe</a>
+            </sgds-link>
+          </sgds-table-cell>
+          <sgds-table-cell>-</sgds-table-cell>
+        </sgds-table-row>
+        <sgds-table-row>
+          <sgds-table-cell>3</sgds-table-cell>
+          <sgds-table-cell>Bob</sgds-table-cell>
+          <sgds-table-cell>Smith</sgds-table-cell>
+          <sgds-table-cell>
+            <sgds-link>
+              <a href="#">@bobsmith</a>
+            </sgds-link>
+          </sgds-table-cell>
+          <sgds-table-cell>
+            <sgds-badge outlined> active </sgds-badge>
+          </sgds-table-cell>
+        </sgds-table-row>
+      </sgds-table>
     </div>
 
     <div class="container">
       <sgds-table
+        headerPosition="vertical"
         responsive="md"
+        columnheader='["#", "First Names", "Last Name", "Username"]'
+        rowheader='["#", "First Names", "Last Name", "Username"]'
+        tabledata='[
+        ["1", "John", "Doe", "@johndoe"],
+        ["2", "Jane", "Doe", "@janedoe"],
+        ["3", "Bob", "Smith", "@bobsmith"]
+        ]'
+      />
+    </div>
+
+    <div class="container">
+      <sgds-table
+        headerPosition="both"
+        responsive="md"
+        columnheader='["First Names", "Last Name", "Username"]'
+        rowheader='["#", "First Names", "Last Name", "Username"]'
+        tabledata='[
+        ["1", "John", "Doe", "@johndoe"],
+        ["2", "Jane", "Doe", "@janedoe"],
+        ["3", "Bob", "Smith", "@bobsmith"]
+        ]'
+        >bye
+        <div>hello</div>
+      </sgds-table>
+    </div>
+
+    <div class="container">
+      <sgds-table
+        headerPosition="horizontal"
+        responsive="md"
+        columnheader='["#", "First Names", "Last Name", "Username"]'
         rowheader='["#", "First Names", "Last Name", "Username"]'
         tabledata='[
         ["1", "John", "Doe", "@johndoe"],
