--- src/components/Table/table-body.css
@@ -1,3 +1,0 @@
-:host {
-  display: table-row-group;
-}
