+++ src/components/Table/table-row.css
@@ -2,8 +2,3 @@
   display: table-row;
   width: 100%;
 }
-
-.header-row::slotted(*) {
-  font-weight: bold;
-  border-bottom: var(--sgds-border-width-1) solid var(--sgds-border-color-emphasis);
-}
